Tank work and move machine blocks under machine folder
