+++ src/main/java/com/veteam/voluminousenergy/blocks/blocks/machines/CarbonShieldedAluminumMachineFrame.java
@@ -1,19 +1,19 @@
-package com.veteam.voluminousenergy.blocks.blocks;
+package com.veteam.voluminousenergy.blocks.blocks.machines;
 
 import com.veteam.voluminousenergy.datagen.VETagDataGenerator;
 import net.minecraft.world.level.block.Block;
 import net.minecraft.world.level.block.SoundType;
 import net.minecraft.world.level.material.Material;
 
-public class TitaniumMachineCasingBlock extends Block {
-    public TitaniumMachineCasingBlock() {
-        super(Block.Properties.of(Material.METAL)
+public class CarbonShieldedAluminumMachineFrame extends Block {
+    public CarbonShieldedAluminumMachineFrame() {
+        super(Properties.of(Material.METAL)
                 .sound(SoundType.METAL)
                 .strength(2.0f)
                 .requiresCorrectToolForDrops()
         );
-        setRegistryName("titanium_machine_casing");
+        setRegistryName("carbon_shielded_aluminum_machine_frame");
         VETagDataGenerator.setRequiresPickaxe(this);
-        VETagDataGenerator.setRequiresDiamond(this);
+        VETagDataGenerator.setRequiresStone(this);
     }
 }
