+++ src/main/java/com/veteam/voluminousenergy/blocks/blocks/machines/CentrifugalAgitatorBlock.java
@@ -1,7 +1,8 @@
-package com.veteam.voluminousenergy.blocks.blocks;
+package com.veteam.voluminousenergy.blocks.blocks.machines;
 
+import com.veteam.voluminousenergy.blocks.blocks.VEBlocks;
 import com.veteam.voluminousenergy.blocks.blocks.util.FaceableBlock;
-import com.veteam.voluminousenergy.blocks.tiles.AqueoulizerTile;
+import com.veteam.voluminousenergy.blocks.tiles.CentrifugalAgitatorTile;
 import com.veteam.voluminousenergy.datagen.VETagDataGenerator;
 import net.minecraft.core.BlockPos;
 import net.minecraft.server.level.ServerPlayer;
@@ -22,16 +23,16 @@ import net.minecraftforge.network.NetworkHooks;
 
 import javax.annotation.Nullable;
 
-public class AqueoulizerBlock extends FaceableBlock implements EntityBlock {
+public class CentrifugalAgitatorBlock extends FaceableBlock implements EntityBlock {
 
-    public AqueoulizerBlock(){
+    public CentrifugalAgitatorBlock(){
         super(Properties.of(Material.STONE)
                 .sound(SoundType.METAL)
                 .strength(2.0f)
                 .lightLevel(l -> 0)
                 .requiresCorrectToolForDrops()
         );
-        setRegistryName("aqueoulizer");
+        setRegistryName("centrifugal_agitator");
         VETagDataGenerator.setRequiresPickaxe(this);
         VETagDataGenerator.setRequiresIron(this);
     }
@@ -39,22 +40,22 @@ public class AqueoulizerBlock extends FaceableBlock implements EntityBlock {
     @Nullable
     @Override
     public BlockEntity newBlockEntity(BlockPos pos, BlockState state) { // Replaces old createBlockEntity method
-        return new AqueoulizerTile(VEBlocks.AQUEOULIZER_TILE, pos, state);
+        return new CentrifugalAgitatorTile(VEBlocks.CENTRIFUGAL_AGITATOR_TILE, pos, state);
     }
 
     // NEW TICK SYSTEM
     @Nullable
-    protected static <T extends BlockEntity> BlockEntityTicker<T> createTicker(Level level, BlockEntityType<T> passedBlockEntity, BlockEntityType<? extends AqueoulizerTile> tile) {
-        return level.isClientSide ? null : createTickerHelper(passedBlockEntity, tile, AqueoulizerTile::serverTick);
+    protected static <T extends BlockEntity> BlockEntityTicker<T> createTicker(Level level, BlockEntityType<T> passedBlockEntity, BlockEntityType<? extends CentrifugalAgitatorTile> tile) {
+        return level.isClientSide ? null : createTickerHelper(passedBlockEntity, tile, CentrifugalAgitatorTile::serverTick);
     }
 
-    public static <T extends BlockEntity, E extends BlockEntity> BlockEntityTicker<T> createTickerHelper(BlockEntityType<T> blockEntityType, BlockEntityType<? extends AqueoulizerTile> tile, BlockEntityTicker<E> serverTick) {
+    public static <T extends BlockEntity, E extends BlockEntity> BlockEntityTicker<T> createTickerHelper(BlockEntityType<T> blockEntityType, BlockEntityType<? extends CentrifugalAgitatorTile> tile, BlockEntityTicker<E> serverTick) {
         return blockEntityType == tile ? (BlockEntityTicker<T>)serverTick : null;
     }
 
     @Nullable
     public <T extends BlockEntity> BlockEntityTicker<T> getTicker(Level level, BlockState state, BlockEntityType<T> blockEntityType) {
-        return createTicker(level, blockEntityType, VEBlocks.AQUEOULIZER_TILE);
+        return createTicker(level, blockEntityType, VEBlocks.CENTRIFUGAL_AGITATOR_TILE);
     }
 
     @Override
@@ -64,7 +65,7 @@ public class AqueoulizerBlock extends FaceableBlock implements EntityBlock {
             if(tileEntity instanceof MenuProvider) {
                 NetworkHooks.openGui((ServerPlayer) player, (MenuProvider) tileEntity, tileEntity.getBlockPos());
             } else {
-                throw new IllegalStateException("Aqueoulizer named container provider is missing!");
+                throw new IllegalStateException("Centrifugal Agitator named container provider is missing!");
             }
             return InteractionResult.SUCCESS;
         }
