+++ src/main/java/com/veteam/voluminousenergy/blocks/blocks/machines/CombustionGeneratorBlock.java
@@ -1,7 +1,8 @@
-package com.veteam.voluminousenergy.blocks.blocks;
+package com.veteam.voluminousenergy.blocks.blocks.machines;
 
+import com.veteam.voluminousenergy.blocks.blocks.VEBlocks;
 import com.veteam.voluminousenergy.blocks.blocks.util.FaceableBlock;
-import com.veteam.voluminousenergy.blocks.tiles.StirlingGeneratorTile;
+import com.veteam.voluminousenergy.blocks.tiles.CombustionGeneratorTile;
 import com.veteam.voluminousenergy.datagen.VETagDataGenerator;
 import net.minecraft.core.BlockPos;
 import net.minecraft.server.level.ServerPlayer;
@@ -22,51 +23,51 @@ import net.minecraftforge.network.NetworkHooks;
 
 import javax.annotation.Nullable;
 
-public class StirlingGeneratorBlock extends FaceableBlock implements EntityBlock {
-    public StirlingGeneratorBlock() {
-        super(Properties.of(Material.METAL)
+public class CombustionGeneratorBlock extends FaceableBlock implements EntityBlock {
+    public CombustionGeneratorBlock(){
+        super(Properties.of(Material.STONE)
                 .sound(SoundType.METAL)
                 .strength(2.0f)
                 .lightLevel(l -> 0)
                 .requiresCorrectToolForDrops()
         );
-        setRegistryName("stirling_generator");
+        setRegistryName("combustion_generator");
         VETagDataGenerator.setRequiresPickaxe(this);
-        VETagDataGenerator.setRequiresStone(this);
+        VETagDataGenerator.setRequiresIron(this);
     }
 
     @Nullable
     @Override
     public BlockEntity newBlockEntity(BlockPos pos, BlockState state) { // Replaces old createBlockEntity method
-        return new StirlingGeneratorTile(VEBlocks.STIRLING_GENERATOR_TILE, pos, state);
+        return new CombustionGeneratorTile(VEBlocks.COMBUSTION_GENERATOR_TILE, pos, state);
     }
 
     // NEW TICK SYSTEM
     @Nullable
-    protected static <T extends BlockEntity> BlockEntityTicker<T> createTicker(Level level, BlockEntityType<T> passedBlockEntity, BlockEntityType<? extends StirlingGeneratorTile> tile) {
-        return level.isClientSide ? null : createTickerHelper(passedBlockEntity, tile, StirlingGeneratorTile::serverTick);
+    protected static <T extends BlockEntity> BlockEntityTicker<T> createTicker(Level level, BlockEntityType<T> passedBlockEntity, BlockEntityType<? extends CombustionGeneratorTile> tile) {
+        return level.isClientSide ? null : createTickerHelper(passedBlockEntity, tile, CombustionGeneratorTile::serverTick);
     }
 
-    public static <T extends BlockEntity, E extends BlockEntity> BlockEntityTicker<T> createTickerHelper(BlockEntityType<T> blockEntityType, BlockEntityType<? extends StirlingGeneratorTile> tile, BlockEntityTicker<E> serverTick) {
+    public static <T extends BlockEntity, E extends BlockEntity> BlockEntityTicker<T> createTickerHelper(BlockEntityType<T> blockEntityType, BlockEntityType<? extends CombustionGeneratorTile> tile, BlockEntityTicker<E> serverTick) {
         return blockEntityType == tile ? (BlockEntityTicker<T>)serverTick : null;
     }
 
     @Nullable
     public <T extends BlockEntity> BlockEntityTicker<T> getTicker(Level level, BlockState state, BlockEntityType<T> blockEntityType) {
-        return createTicker(level, blockEntityType, VEBlocks.STIRLING_GENERATOR_TILE);
+        return createTicker(level, blockEntityType, VEBlocks.COMBUSTION_GENERATOR_TILE);
     }
+
     @Override
     public InteractionResult use(BlockState state, Level world, BlockPos pos, Player player, InteractionHand handIn, BlockHitResult hit){
-        if(!world.isClientSide) {
+        if (!world.isClientSide) {
             BlockEntity tileEntity = world.getBlockEntity(pos);
-            if(tileEntity instanceof MenuProvider) {
+            if (tileEntity instanceof MenuProvider) {
                 NetworkHooks.openGui((ServerPlayer) player, (MenuProvider) tileEntity, tileEntity.getBlockPos());
             } else {
-                throw new IllegalStateException("Stirling Generator named container provider is missing!");
+                throw new IllegalStateException("Combustion Generator named container provider is missing!");
             }
-            return InteractionResult.SUCCESS;
         }
         return InteractionResult.SUCCESS;
-
     }
+
 }
