+++ src/main/java/com/veteam/voluminousenergy/blocks/blocks/machines/PrimitiveBlastFurnaceBlock.java
@@ -1,6 +1,7 @@
-package com.veteam.voluminousenergy.blocks.blocks;
+package com.veteam.voluminousenergy.blocks.blocks.machines;
 
 
+import com.veteam.voluminousenergy.blocks.blocks.VEBlocks;
 import com.veteam.voluminousenergy.blocks.blocks.util.FaceableBlock;
 import com.veteam.voluminousenergy.blocks.tiles.PrimitiveBlastFurnaceTile;
 import com.veteam.voluminousenergy.datagen.VETagDataGenerator;
