+++ src/main/java/com/veteam/voluminousenergy/blocks/blocks/machines/PrimitiveStirlingGeneratorBlock.java
@@ -1,7 +1,8 @@
-package com.veteam.voluminousenergy.blocks.blocks;
+package com.veteam.voluminousenergy.blocks.blocks.machines;
 
+import com.veteam.voluminousenergy.blocks.blocks.VEBlocks;
 import com.veteam.voluminousenergy.blocks.blocks.util.FaceableBlock;
-import com.veteam.voluminousenergy.blocks.tiles.BlastFurnaceTile;
+import com.veteam.voluminousenergy.blocks.tiles.PrimitiveStirlingGeneratorTile;
 import com.veteam.voluminousenergy.datagen.VETagDataGenerator;
 import net.minecraft.core.BlockPos;
 import net.minecraft.server.level.ServerPlayer;
@@ -22,49 +23,50 @@ import net.minecraftforge.network.NetworkHooks;
 
 import javax.annotation.Nullable;
 
-public class BlastFurnaceBlock extends FaceableBlock implements EntityBlock {
+public class PrimitiveStirlingGeneratorBlock extends FaceableBlock implements EntityBlock {
 
-    public BlastFurnaceBlock(){
+    public PrimitiveStirlingGeneratorBlock() {
+
         super(Properties.of(Material.STONE)
-                .sound(SoundType.METAL)
-                .strength(3.0f)
+                .sound(SoundType.STONE)
+                .strength(2.0f)
                 .lightLevel(l -> 0)
                 .requiresCorrectToolForDrops()
         );
-        setRegistryName("blast_furnace");
+
+        setRegistryName("primitivestirlinggenerator");
         VETagDataGenerator.setRequiresPickaxe(this);
-        VETagDataGenerator.setRequiresDiamond(this);
+        VETagDataGenerator.setRequiresWood(this);
     }
 
-    @Nullable
     @Override
     public BlockEntity newBlockEntity(BlockPos pos, BlockState state) { // Replaces old createBlockEntity method
-        return new BlastFurnaceTile(VEBlocks.BLAST_FURNACE_TILE, pos, state);
+        return new PrimitiveStirlingGeneratorTile(VEBlocks.PRIMITIVE_STIRLING_GENERATOR_TILE, pos, state);
     }
 
     // NEW TICK SYSTEM
     @Nullable
-    protected static <T extends BlockEntity> BlockEntityTicker<T> createTicker(Level level, BlockEntityType<T> passedBlockEntity, BlockEntityType<? extends BlastFurnaceTile> tile) {
-        return level.isClientSide ? null : createTickerHelper(passedBlockEntity, tile, BlastFurnaceTile::serverTick);
+    protected static <T extends BlockEntity> BlockEntityTicker<T> createTicker(Level level, BlockEntityType<T> passedBlockEntity, BlockEntityType<? extends PrimitiveStirlingGeneratorTile> tile) {
+        return level.isClientSide ? null : createTickerHelper(passedBlockEntity, tile, PrimitiveStirlingGeneratorTile::serverTick);
     }
 
-    public static <T extends BlockEntity, E extends BlockEntity> BlockEntityTicker<T> createTickerHelper(BlockEntityType<T> blockEntityType, BlockEntityType<? extends BlastFurnaceTile> tile, BlockEntityTicker<E> serverTick) {
+    public static <T extends BlockEntity, E extends BlockEntity> BlockEntityTicker<T> createTickerHelper(BlockEntityType<T> blockEntityType, BlockEntityType<? extends PrimitiveStirlingGeneratorTile> tile, BlockEntityTicker<E> serverTick) {
         return blockEntityType == tile ? (BlockEntityTicker<T>)serverTick : null;
     }
 
     @Nullable
     public <T extends BlockEntity> BlockEntityTicker<T> getTicker(Level level, BlockState state, BlockEntityType<T> blockEntityType) {
-        return createTicker(level, blockEntityType, VEBlocks.BLAST_FURNACE_TILE);
+        return createTicker(level, blockEntityType, VEBlocks.PRIMITIVE_STIRLING_GENERATOR_TILE);
     }
 
     @Override
-    public InteractionResult use(BlockState state, Level world, BlockPos pos, Player player, InteractionHand handIn, BlockHitResult hit){
-        if(!world.isClientSide) {
+    public InteractionResult use(BlockState state, Level world, BlockPos pos, Player player, InteractionHand hand, BlockHitResult result) {
+        if (!world.isClientSide){
             BlockEntity tileEntity = world.getBlockEntity(pos);
-            if(tileEntity instanceof MenuProvider) {
+            if (tileEntity instanceof MenuProvider){
                 NetworkHooks.openGui((ServerPlayer) player, (MenuProvider) tileEntity, tileEntity.getBlockPos());
             } else {
-                throw new IllegalStateException("Blast Furnace named container provider is missing!");
+                throw new IllegalStateException("Primitive Stirling named container provider is missing!");
             }
             return InteractionResult.SUCCESS;
         }
