+++ src/main/java/com/veteam/voluminousenergy/blocks/blocks/machines/SolarPanelBlock.java
@@ -1,7 +1,8 @@
-package com.veteam.voluminousenergy.blocks.blocks;
+package com.veteam.voluminousenergy.blocks.blocks.machines;
 
+import com.veteam.voluminousenergy.blocks.blocks.VEBlocks;
 import com.veteam.voluminousenergy.blocks.blocks.util.FaceableBlock;
-import com.veteam.voluminousenergy.blocks.tiles.DistillationUnitTile;
+import com.veteam.voluminousenergy.blocks.tiles.SolarPanelTile;
 import com.veteam.voluminousenergy.datagen.VETagDataGenerator;
 import net.minecraft.core.BlockPos;
 import net.minecraft.server.level.ServerPlayer;
@@ -22,41 +23,39 @@ import net.minecraftforge.network.NetworkHooks;
 
 import javax.annotation.Nullable;
 
-public class DistillationUnitBlock extends FaceableBlock implements EntityBlock {
-
-    public DistillationUnitBlock(){
-        super(Properties.of(Material.STONE)
+public class SolarPanelBlock extends FaceableBlock implements EntityBlock {
+    public SolarPanelBlock() {
+        super(Properties.of(Material.METAL)
                 .sound(SoundType.METAL)
                 .strength(2.0f)
                 .lightLevel(l -> 0)
                 .requiresCorrectToolForDrops()
         );
-        setRegistryName("distillation_unit");
+        setRegistryName("solar_panel");
         VETagDataGenerator.setRequiresPickaxe(this);
-        VETagDataGenerator.setRequiresIron(this);
+        VETagDataGenerator.setRequiresStone(this);
     }
 
     @Nullable
     @Override
     public BlockEntity newBlockEntity(BlockPos pos, BlockState state) { // Replaces old createBlockEntity method
-        return new DistillationUnitTile(VEBlocks.DISTILLATION_UNIT_TILE, pos, state);
+        return new SolarPanelTile(VEBlocks.SOLAR_PANEL_TILE, pos, state);
     }
 
     // NEW TICK SYSTEM
     @Nullable
-    protected static <T extends BlockEntity> BlockEntityTicker<T> createTicker(Level level, BlockEntityType<T> passedBlockEntity, BlockEntityType<? extends DistillationUnitTile> tile) {
-        return level.isClientSide ? null : createTickerHelper(passedBlockEntity, tile, DistillationUnitTile::serverTick);
+    protected static <T extends BlockEntity> BlockEntityTicker<T> createTicker(Level level, BlockEntityType<T> passedBlockEntity, BlockEntityType<? extends SolarPanelTile> tile) {
+        return level.isClientSide ? null : createTickerHelper(passedBlockEntity, tile, SolarPanelTile::serverTick);
     }
 
-    public static <T extends BlockEntity, E extends BlockEntity> BlockEntityTicker<T> createTickerHelper(BlockEntityType<T> blockEntityType, BlockEntityType<? extends DistillationUnitTile> tile, BlockEntityTicker<E> serverTick) {
+    public static <T extends BlockEntity, E extends BlockEntity> BlockEntityTicker<T> createTickerHelper(BlockEntityType<T> blockEntityType, BlockEntityType<? extends SolarPanelTile> tile, BlockEntityTicker<E> serverTick) {
         return blockEntityType == tile ? (BlockEntityTicker<T>)serverTick : null;
     }
 
     @Nullable
     public <T extends BlockEntity> BlockEntityTicker<T> getTicker(Level level, BlockState state, BlockEntityType<T> blockEntityType) {
-        return createTicker(level, blockEntityType, VEBlocks.DISTILLATION_UNIT_TILE);
+        return createTicker(level, blockEntityType, VEBlocks.SOLAR_PANEL_TILE);
     }
-
     @Override
     public InteractionResult use(BlockState state, Level world, BlockPos pos, Player player, InteractionHand handIn, BlockHitResult hit){
         if(!world.isClientSide) {
@@ -64,7 +63,7 @@ public class DistillationUnitBlock extends FaceableBlock implements EntityBlock 
             if(tileEntity instanceof MenuProvider) {
                 NetworkHooks.openGui((ServerPlayer) player, (MenuProvider) tileEntity, tileEntity.getBlockPos());
             } else {
-                throw new IllegalStateException("Distillation Unit named container provider is missing!");
+                throw new IllegalStateException("Solar Panel named container provider is missing!");
             }
             return InteractionResult.SUCCESS;
         }
