+++ src/main/java/com/veteam/voluminousenergy/blocks/blocks/machines/StirlingGeneratorBlock.java
@@ -1,7 +1,8 @@
-package com.veteam.voluminousenergy.blocks.blocks;
+package com.veteam.voluminousenergy.blocks.blocks.machines;
 
+import com.veteam.voluminousenergy.blocks.blocks.VEBlocks;
 import com.veteam.voluminousenergy.blocks.blocks.util.FaceableBlock;
-import com.veteam.voluminousenergy.blocks.tiles.CentrifugalAgitatorTile;
+import com.veteam.voluminousenergy.blocks.tiles.StirlingGeneratorTile;
 import com.veteam.voluminousenergy.datagen.VETagDataGenerator;
 import net.minecraft.core.BlockPos;
 import net.minecraft.server.level.ServerPlayer;
@@ -22,41 +23,39 @@ import net.minecraftforge.network.NetworkHooks;
 
 import javax.annotation.Nullable;
 
-public class CentrifugalAgitatorBlock extends FaceableBlock implements EntityBlock {
-
-    public CentrifugalAgitatorBlock(){
-        super(Properties.of(Material.STONE)
+public class StirlingGeneratorBlock extends FaceableBlock implements EntityBlock {
+    public StirlingGeneratorBlock() {
+        super(Properties.of(Material.METAL)
                 .sound(SoundType.METAL)
                 .strength(2.0f)
                 .lightLevel(l -> 0)
                 .requiresCorrectToolForDrops()
         );
-        setRegistryName("centrifugal_agitator");
+        setRegistryName("stirling_generator");
         VETagDataGenerator.setRequiresPickaxe(this);
-        VETagDataGenerator.setRequiresIron(this);
+        VETagDataGenerator.setRequiresStone(this);
     }
 
     @Nullable
     @Override
     public BlockEntity newBlockEntity(BlockPos pos, BlockState state) { // Replaces old createBlockEntity method
-        return new CentrifugalAgitatorTile(VEBlocks.CENTRIFUGAL_AGITATOR_TILE, pos, state);
+        return new StirlingGeneratorTile(VEBlocks.STIRLING_GENERATOR_TILE, pos, state);
     }
 
     // NEW TICK SYSTEM
     @Nullable
-    protected static <T extends BlockEntity> BlockEntityTicker<T> createTicker(Level level, BlockEntityType<T> passedBlockEntity, BlockEntityType<? extends CentrifugalAgitatorTile> tile) {
-        return level.isClientSide ? null : createTickerHelper(passedBlockEntity, tile, CentrifugalAgitatorTile::serverTick);
+    protected static <T extends BlockEntity> BlockEntityTicker<T> createTicker(Level level, BlockEntityType<T> passedBlockEntity, BlockEntityType<? extends StirlingGeneratorTile> tile) {
+        return level.isClientSide ? null : createTickerHelper(passedBlockEntity, tile, StirlingGeneratorTile::serverTick);
     }
 
-    public static <T extends BlockEntity, E extends BlockEntity> BlockEntityTicker<T> createTickerHelper(BlockEntityType<T> blockEntityType, BlockEntityType<? extends CentrifugalAgitatorTile> tile, BlockEntityTicker<E> serverTick) {
+    public static <T extends BlockEntity, E extends BlockEntity> BlockEntityTicker<T> createTickerHelper(BlockEntityType<T> blockEntityType, BlockEntityType<? extends StirlingGeneratorTile> tile, BlockEntityTicker<E> serverTick) {
         return blockEntityType == tile ? (BlockEntityTicker<T>)serverTick : null;
     }
 
     @Nullable
     public <T extends BlockEntity> BlockEntityTicker<T> getTicker(Level level, BlockState state, BlockEntityType<T> blockEntityType) {
-        return createTicker(level, blockEntityType, VEBlocks.CENTRIFUGAL_AGITATOR_TILE);
+        return createTicker(level, blockEntityType, VEBlocks.STIRLING_GENERATOR_TILE);
     }
-
     @Override
     public InteractionResult use(BlockState state, Level world, BlockPos pos, Player player, InteractionHand handIn, BlockHitResult hit){
         if(!world.isClientSide) {
@@ -64,10 +63,11 @@ public class CentrifugalAgitatorBlock extends FaceableBlock implements EntityBlo
             if(tileEntity instanceof MenuProvider) {
                 NetworkHooks.openGui((ServerPlayer) player, (MenuProvider) tileEntity, tileEntity.getBlockPos());
             } else {
-                throw new IllegalStateException("Centrifugal Agitator named container provider is missing!");
+                throw new IllegalStateException("Stirling Generator named container provider is missing!");
             }
             return InteractionResult.SUCCESS;
         }
         return InteractionResult.SUCCESS;
+
     }
 }
