+++ src/main/java/com/veteam/voluminousenergy/blocks/blocks/machines/ToolingStationBlock.java
@@ -1,7 +1,8 @@
-package com.veteam.voluminousenergy.blocks.blocks;
+package com.veteam.voluminousenergy.blocks.blocks.machines;
 
+import com.veteam.voluminousenergy.blocks.blocks.VEBlocks;
 import com.veteam.voluminousenergy.blocks.blocks.util.FaceableBlock;
-import com.veteam.voluminousenergy.blocks.tiles.ElectrolyzerTile;
+import com.veteam.voluminousenergy.blocks.tiles.ToolingStationTile;
 import com.veteam.voluminousenergy.datagen.VETagDataGenerator;
 import net.minecraft.core.BlockPos;
 import net.minecraft.server.level.ServerPlayer;
@@ -22,40 +23,41 @@ import net.minecraftforge.network.NetworkHooks;
 
 import javax.annotation.Nullable;
 
-public class ElectrolyzerBlock extends FaceableBlock implements EntityBlock {
+public class ToolingStationBlock extends FaceableBlock implements EntityBlock {
 
-    public ElectrolyzerBlock() {
-        super(Properties.of(Material.METAL)
+    public ToolingStationBlock(){
+        super(Properties.of(Material.STONE)
                 .sound(SoundType.METAL)
                 .strength(2.0f)
                 .lightLevel(l -> 0)
                 .requiresCorrectToolForDrops()
         );
-        setRegistryName("electrolyzer");
+        setRegistryName("tooling_station");
         VETagDataGenerator.setRequiresPickaxe(this);
-        VETagDataGenerator.setRequiresStone(this);
+        VETagDataGenerator.setRequiresDiamond(this);
     }
 
     @Nullable
     @Override
     public BlockEntity newBlockEntity(BlockPos pos, BlockState state) { // Replaces old createBlockEntity method
-        return new ElectrolyzerTile(VEBlocks.ELECTROLYZER_TILE, pos, state);
+        return new ToolingStationTile(VEBlocks.TOOLING_STATION_TILE, pos, state);
     }
 
     // NEW TICK SYSTEM
     @Nullable
-    protected static <T extends BlockEntity> BlockEntityTicker<T> createTicker(Level level, BlockEntityType<T> passedBlockEntity, BlockEntityType<? extends ElectrolyzerTile> tile) {
-        return level.isClientSide ? null : createTickerHelper(passedBlockEntity, tile, ElectrolyzerTile::serverTick);
+    protected static <T extends BlockEntity> BlockEntityTicker<T> createTicker(Level level, BlockEntityType<T> passedBlockEntity, BlockEntityType<? extends ToolingStationTile> tile) {
+        return level.isClientSide ? null : createTickerHelper(passedBlockEntity, tile, ToolingStationTile::serverTick);
     }
 
-    public static <T extends BlockEntity, E extends BlockEntity> BlockEntityTicker<T> createTickerHelper(BlockEntityType<T> blockEntityType, BlockEntityType<? extends ElectrolyzerTile> tile, BlockEntityTicker<E> serverTick) {
+    public static <T extends BlockEntity, E extends BlockEntity> BlockEntityTicker<T> createTickerHelper(BlockEntityType<T> blockEntityType, BlockEntityType<? extends ToolingStationTile> tile, BlockEntityTicker<E> serverTick) {
         return blockEntityType == tile ? (BlockEntityTicker<T>)serverTick : null;
     }
 
     @Nullable
     public <T extends BlockEntity> BlockEntityTicker<T> getTicker(Level level, BlockState state, BlockEntityType<T> blockEntityType) {
-        return createTicker(level, blockEntityType, VEBlocks.ELECTROLYZER_TILE);
+        return createTicker(level, blockEntityType, VEBlocks.TOOLING_STATION_TILE);
     }
+
     @Override
     public InteractionResult use(BlockState state, Level world, BlockPos pos, Player player, InteractionHand handIn, BlockHitResult hit){
         if(!world.isClientSide) {
@@ -63,12 +65,10 @@ public class ElectrolyzerBlock extends FaceableBlock implements EntityBlock {
             if(tileEntity instanceof MenuProvider) {
                 NetworkHooks.openGui((ServerPlayer) player, (MenuProvider) tileEntity, tileEntity.getBlockPos());
             } else {
-                throw new IllegalStateException("Electrolyzer named container provider is missing!");
+                throw new IllegalStateException("Tooling Station named container provider is missing!");
             }
             return InteractionResult.SUCCESS;
         }
         return InteractionResult.SUCCESS;
-
     }
-
 }
