+++ src/main/java/com/veteam/voluminousenergy/blocks/blocks/machines/tanks/TankBlock.java
@@ -1,0 +1,43 @@
+package com.veteam.voluminousenergy.blocks.blocks.machines.tanks;
+
+import com.veteam.voluminousenergy.blocks.blocks.VEBlocks;
+import com.veteam.voluminousenergy.blocks.blocks.util.FaceableBlock;
+import com.veteam.voluminousenergy.blocks.tiles.TankTile;
+import com.veteam.voluminousenergy.datagen.VETagDataGenerator;
+import net.minecraft.core.BlockPos;
+import net.minecraft.server.level.ServerPlayer;
+import net.minecraft.world.InteractionHand;
+import net.minecraft.world.InteractionResult;
+import net.minecraft.world.MenuProvider;
+import net.minecraft.world.entity.player.Player;
+import net.minecraft.world.level.Level;
+import net.minecraft.world.level.block.EntityBlock;
+import net.minecraft.world.level.block.SoundType;
+import net.minecraft.world.level.block.entity.BlockEntity;
+import net.minecraft.world.level.block.entity.BlockEntityTicker;
+import net.minecraft.world.level.block.entity.BlockEntityType;
+import net.minecraft.world.level.block.state.BlockState;
+import net.minecraft.world.level.material.Material;
+import net.minecraft.world.phys.BlockHitResult;
+import net.minecraftforge.network.NetworkHooks;
+
+import javax.annotation.Nullable;
+
+public class TankBlock extends FaceableBlock {
+
+    public TankBlock(Properties properties){ super(properties); }
+
+    @Override
+    public InteractionResult use(BlockState state, Level world, BlockPos pos, Player player, InteractionHand handIn, BlockHitResult hit){
+        if(!world.isClientSide) {
+            BlockEntity tileEntity = world.getBlockEntity(pos);
+            if(tileEntity instanceof MenuProvider) {
+                NetworkHooks.openGui((ServerPlayer) player, (MenuProvider) tileEntity, tileEntity.getBlockPos());
+            } else {
+                throw new IllegalStateException("TankBlock named container provider is missing!");
+            }
+            return InteractionResult.SUCCESS;
+        }
+        return InteractionResult.SUCCESS;
+    }
+}
