+++ src/main/java/com/veteam/voluminousenergy/blocks/containers/TankContainer.java
@@ -83,7 +83,7 @@ public class TankContainer extends VoluminousContainer {
     public void updateDirectionButton(int direction, int slotId){ this.screen.updateButtonDirection(direction,slotId); }
 
     @Override
-public void updateStatusButton(boolean status, int slotId){
+    public void updateStatusButton(boolean status, int slotId){
         this.screen.updateBooleanButton(status, slotId);
     }
 }
